refactor(stats): use dynamic import for subcommand
