+++ packages/opencode/src/index.ts
@@ -17,7 +17,7 @@ import { FormatError } from "./cli/error"
 import { ServeCommand } from "./cli/cmd/serve/command"
 import { Filesystem } from "@/util/filesystem"
 import { DebugCommand } from "./cli/cmd/debug/command"
-import { StatsCommand } from "./cli/cmd/stats"
+import { StatsCommand } from "./cli/cmd/stats/command"
 import { McpCommand } from "./cli/cmd/mcp/command"
 import { GithubCommand } from "./cli/cmd/github/command"
 import { ExportCommand } from "./cli/cmd/export/command"
